// resume-web/src/app/shared/models/add-education.model.ts
/**
 * Represents an add education model
 */
export interface AddEducationModel {

  /**
   * Gets or sets the use identifier
   */
  userId: string;

  /**
   * Gets or sets the school
   */
  school: string;

  /**
   * Gets or sets the degree
   */
  degree: string;

  /**
   * Gets or sets the date from
   */
  from: string;

  /**
   * Gets or sets the to
   */
  to: string;
}